MINIFICPP-2110 Fix default CA path for S3 on CentOS
