+++ extensions/http-curl/client/HTTPClient.h
@@ -188,7 +188,6 @@ class HTTPClient : public utils::BaseHTTPClient, public core::Connectable {
 
   static bool isValidHttpHeaderField(std::string_view field_name);
   static std::string replaceInvalidCharactersInHttpHeaderFieldName(std::string field_name);
-  static std::optional<std::filesystem::path> getDefaultCAPath();
 
  private:
   static int onProgress(void *client, curl_off_t dltotal, curl_off_t dlnow, curl_off_t ultotal, curl_off_t ulnow);
