+++ extensions/http-curl/tests/C2NullConfiguration.cpp
@@ -75,7 +75,7 @@ class VerifyC2Server : public HTTPIntegrationBase {
     std::string port;
     std::string scheme;
     std::string path;
-    parse_http_components(url, port, scheme, path);
+    minifi::utils::parse_http_components(url, port, scheme, path);
     configuration->set(Configuration::nifi_c2_enable, "true");
     configuration->set(Configuration::nifi_c2_agent_class, "test");
     configuration->set(Configuration::nifi_c2_agent_protocol_class, "RESTSender");
